+++ src/JK.Common/CHANGELOG.md
@@ -15,6 +15,7 @@ Unless otherwise noted, all changes made by [@jeremyknight-me](https://github.co
 - Added Singleton pattern Lazy<T> implementation reference class
 - Added DateOnly version of DateHelper methods. 
 - Added DateOnly extension methods.
+- Added DecimalHelper with `GetDecimalPart` and `GetWholePart` methods
 
 ### Changed